backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
docling==2.31.0
docling-core==2.28.1
docling-ibm-models==3.4.2
docling-parse==4.0.1
docopt==0.6.2
easyocr==1.7.2
effdet==0.4.1
elevenlabs==1.54.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.115.11
fastembed==0.6.1
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.1
google_search_results==2.4.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
groovy==0.1.2
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.44
langchain-openai==0.3.8
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.6
langdetect==1.0.9
langgraph==0.3.9
langgraph-checkpoint==2.0.19
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.13
latex2mathml==3.77.0
lazy_loader==0.4
loguru==0.7.3
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.18
mypy-extensions==1.0.0
narwhals==1.30.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nilearn==0.11.1
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
openai==1.66.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20240706
pi_heif==0.22.0
pickleshare==0.7.5
pikepdf==9.5.2
pillow==11.1.0
pip==25.0
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.1
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
pytest==8.3.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
qdrant-client==1.13.3
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
rtree==1.4.0
ruff==0.10.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
semchunk==2.2.2
sentence-transformers==3.4.1
setuptools==75.8.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
stack_data==0.6.3
starlette==0.46.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tifffile==2025.2.18
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.1
urllib3==1.26.20
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0